+++ src/Screenshot.h
@@ -4,5 +4,6 @@
 void TakeScreenshots();
 void RegisterScreenshotHotkey(HWND hwnd);
 void UnregisterScreenshotHotkey(HWND hwnd);
+void ShowSetScreenshotHotkeyDialog(HWND hwndOwner);
 
 constexpr int kScreenshotHotkeyId = 0x5001;